feat: initial refactor. Still missing unit testing on new functions.
