+++ src/ansys/mapdl/core/__init__.py
@@ -110,7 +110,9 @@ from ansys.mapdl.core.launcher import close_all_local_instances
 
 # override default launcher when on pyansys.com
 if "ANSJUPHUB_VER" in os.environ:  # pragma: no cover
-    from ansys.mapdl.core.jupyter import launch_mapdl_on_cluster as launch_mapdl
+    from ansys.mapdl.core.launcher.jupyter import (
+        launch_mapdl_on_cluster as launch_mapdl,
+    )
 else:
     from ansys.mapdl.core.launcher import launch_mapdl
