+++ src/ansys/mapdl/core/launcher/jupyter.py
