+++ tests/test_launcher/test_console.py
@@ -1,0 +1,50 @@
+# Copyright (C) 2016 - 2025 ANSYS, Inc. and/or its affiliates.
+# SPDX-License-Identifier: MIT
+#
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+
+from ansys.mapdl.core.launcher.console import launch_mapdl_console
+from conftest import requires
+
+
+@requires("console")
+def test_launch_mapdl_console(tmpdir):
+    filename = str(tmpdir.mkdir("tmpdir").join("tmp.inp"))
+
+    mapdl = launch_mapdl_console(log_apdl=filename, mode="console")
+
+    mapdl.prep7()
+    mapdl.run("!comment test")
+    mapdl.k(1, 0, 0, 0)
+    mapdl.k(2, 1, 0, 0)
+    mapdl.k(3, 1, 1, 0)
+    mapdl.k(4, 0, 1, 0)
+
+    mapdl.exit()
+
+    with open(filename, "r") as fid:
+        text = "".join(fid.readlines())
+
+    assert "PREP7" in text
+    assert "!comment test" in text
+    assert "K,1,0,0,0" in text
+    assert "K,2,1,0,0" in text
+    assert "K,3,1,1,0" in text
+    assert "K,4,0,1,0" in text
